discord-bot/requirements.txt
discord.py
aiohttp
redis
python-dotenv